==> code/01-eda-jobs-data.ipynb <==
# EDA

import pandas as pd
import numpy as np
import re

import httpimport
url = 'https://raw.githubusercontent.com/zach-brown-18/class-toolkit/main/eda/'
with httpimport.remote_repo(['cleaning', 'columns'], url):
    import cleaning as c
    import columns as cols

---

# Load and Look at Data

dice = pd.read_csv('../data/dice.csv', encoding='latin-1')
print(dice.shape)

# Drop the fluff
use_cols = ['date_added', 'job_description', 'job_title', 'job_type', 'location', 'organization', 'sector']
dice = dice.loc[:, use_cols]

# 7 total columns used
print(dice.shape)
dice.head(2)

---

# Cleaning

## Functions

def remove_duplicate_skills(idx,df):
    skills = df.loc[idx, 'sector'].split(' -')[0]
    try:
        df.loc[idx, 'job_description'] = df.loc[idx, 'job_description'].split(skills)[1]
    except:
        print(f'{idx} failed.')

def remove_unwanted_characters(s):
    s = re.sub('\*', ' ', s)
    s = re.sub('/', ' ', s)
    s = re.sub('_', ' ', s)
    s = re.sub('\?', ' ', s)
    s = re.sub('%', ' ', s)
    s = re.sub('@', ' ', s)
    s = re.sub('#', ' ', s)
    s = re.sub('!', ' ', s)
    s = re.sub(',', ' ', s)
    s = re.sub('\+', ' ', s)
    s = re.sub('&', ' ', s)
    s = re.sub('\d', ' ', s)
    return s

def rem_fluff(data):
    data = data.replace('please share resume', " ")
    data = data.replace('call me', " ")
    data = data.replace('if you are a', " ")
    data = data.replace('if you are an', " ")
    data = data.replace('job description', " ")
    data = data.replace('if you are', " ")
    data = data.replace('reply', " ")
    data = data.replace("please apply today applicants must be authorized to work in the u.s.please apply directly to by clicking 'click here to apply' with your word resume looking forward to receiving your resume and going over the position in more detail with you.- not a fit for this position click the link at the bottom of this email to search all of our open positions.looking forward to receiving your resume cybercoderscybercoders inc is proud to be an equal opportunity employerall qualified applicants will receive consideration for employment without regard to race color religion sex national origin disability protected veteran status or any other characteristic protected by law.your right to work - in compliance with federal law all persons hired will be required to verify identity and eligibility to work in the united states and to complete the required employment eligibility verification document form upon hire.copyright  - . cybercoders inc. all rights reserved.", ' ')
    return data

# function to get rid of repeated skills after a '-'
def split_skills(data):
        x = data
        nospace = x.replace(' ', '') # removing spaces
        if '-' in nospace:
            first_half = nospace.split('-')[0] #first half of string before '-'
            second_half = nospace.split('-')[1] # second half of string after '-'
            if first_half == second_half:
                x = list(x.split('-'))[0].strip() #getting only first half of string to return
                return x
            else:
                return x      
        else:
            return x

def drop_corrupt_rows(df):
    df['word_count'] = df['job_description'].map(lambda x: len(x.split()))
    
    mean_word_count = df['word_count'].mean()
    print('Mean word count per post:', round(mean_word_count))
    
    std_word_count = df['word_count'].std()
    print('std word count per post:', round(std_word_count))
    
    too_short = mean_word_count - 1.5*std_word_count
    too_short = 20
    
    # Drop posts with 20 word or less
    print(len(df), 'before dropping')
    mask = df['word_count'] > too_short
    df = df[mask]
    print(len(df), 'after dropping')

    df.drop(columns='word_count', inplace=True)
    
    return df.reset_index(drop=True)

---

## Cleaning job_description

# Remove unwanted characters and numbers
dice['sector'].fillna('', inplace=True)
dice['job_description'] = dice['job_description'].map(remove_unwanted_characters)
dice['sector'] = dice['sector'].map(remove_unwanted_characters)

# Correct multiple spaces (including \t and \n)
dice['job_description'] = dice['job_description'].map(lambda x: re.sub('\s+',' ', x))
dice['sector'].fillna('', inplace=True)
dice['sector'] = dice['sector'].map(lambda x: re.sub('\s+',' ', x))

# Remove special characters
dice['job_description'] = dice['job_description'].map(c.remove_special_chars)
dice['sector'] = dice['sector'].map(c.remove_special_chars)

# Drop corrupted rows
dice = drop_corrupt_rows(dice)

# Remove skills repeated in job description and sector
skills_repeated = dice['job_description'].map(lambda x: 'Minimum Required Skills' in x)
for row in dice.loc[skills_repeated, :].index:
    remove_duplicate_skills(row, dice)
    
# Lowercase job_description
dice['job_description'] = dice['job_description'].str.lower()

# Remove common phrases
dice['job_description'] = dice['job_description'].map(rem_fluff)
dice['sector'] = dice['sector'].map(rem_fluff)

# Correct multiple spaces (including \t and \n)
dice['job_description'] = dice['job_description'].map(lambda x: re.sub('\s+',' ', x))
dice['sector'] = dice['sector'].map(lambda x: re.sub('\s+',' ', x))

## Search Duplicate Posts

# Find duplicate posts
description_counts = dice['job_description'].value_counts()
multiple_posts = description_counts > 1
description_counts[multiple_posts].sum()

# CyberCoders are spamming the list
repeated_descriptions = description_counts[multiple_posts].index
repeated_descriptions = dice['job_description'].map(lambda x: x in repeated_descriptions)

dice.loc[repeated_descriptions, :]['organization'].value_counts().head()

# Majority of repeated posts are only repeated once
description_counts[description_counts == 2].sum()

**Conclusion:** Repeated, identical posts will distort the data pool. If a post is repeated, drop all occurances but one.

## Drop Duplicate job posts

print(f'{dice.shape[0]} job listings before dropping duplicates')

# Define duplicate posts as having identical job description, job type and organization.
non_duplicates = dice.drop(columns=['date_added', 'location', 'job_title', 'sector']).drop_duplicates().index
dice = dice.loc[non_duplicates, :].reset_index(drop=True)

print(f'{dice.shape[0]} job listings after dropping duplicates')

# Find duplicate posts - still some left over
description_counts = dice['job_description'].value_counts()
multiple_posts = description_counts > 1
print(f'{description_counts[multiple_posts].sum()} duplicates left over')

## Remove phone numbers and email addresses

# Remove them from the job description
dice['job_description'] = dice['job_description'].map(c.remove_phone_numbers)
dice['job_description'] = dice['job_description'].map(c.remove_emails)

dice.reset_index(drop=True, inplace=True)

## Cleaning job_title

# Finding null values in job_title
dice[dice['job_title'].isnull()]

# Removing the one null job_title
dice = dice.dropna(subset = ['job_title'])

### catagorizing job_title

dice.loc[dice['job_title'].str.contains('Developer|developer|Dev|dev|Scrum|scrum'), 
                  'job_title'] = 'Developer'
dice.loc[dice['job_title'].str.contains('Analyst|analyst'), 
                  'job_title'] = 'Analyst'
dice.loc[dice['job_title'].str.contains('Programmer|programmer|Programming|programming|Full'),
                  'job_title'] = 'Programmer'
dice.loc[dice['job_title'].str.contains('Manager|manager|Project Coordinator|Technical Lead'), 
                  'job_title'] = 'Manager'
dice.loc[dice['job_title'].str.contains('Engineer|engineer'), 
                  'job_title'] = 'Engineer'
dice.loc[dice['job_title'].str.contains('Architect|architect'), 
                  'job_title'] = 'Architect'
dice.loc[dice['job_title'].str.contains('Designer|designer'), 
                  'job_title'] = 'Designer'
dice.loc[dice['job_title'].str.contains('Technician|technician'), 
                  'job_title'] = 'Technician'
dice.loc[dice['job_title'].str.contains('Administrator|administrator|Admin|admin'), 
                  'job_title'] = 'Administrator'
dice.loc[dice['job_title'].str.contains('Consulting|consulting|Consultation|consultation|Consultant|consultant'), 
                  'job_title'] = 'Consulting'
dice.loc[dice['job_title'].str.contains('Support|support|Helpdesk'), 
                  'job_title'] = 'Support'
dice.loc[dice['job_title'].str.contains('Director|director|CTO'),
                  'job_title'] = 'Director'
dice.loc[dice['job_title'].str.contains('Entry'), 
                  'job_title'] = 'Entry Position'
dice.loc[dice['job_title'].str.contains('Data Scientist|SQL DBA|SQL Server DBA|IT'), 
                  'job_title'] = 'Data Position'

dice['job_title'].value_counts()[0:13].sum()

# Creating a list of the top 13 most popular job_titles
keep_titles = dice['job_title'].value_counts()[:13].index.tolist()
# Only keeping the top 13 job_titles from keep_titles
dice = dice[dice['job_title'].isin(keep_titles)]

dice['job_title'].value_counts()

## Cleaning location

### Catagorizing location

dice.loc[dice['location'].str.contains('ME|NH|VT|MA|MARYLAND|Laurel|CT|Ct|RI|NY|Ny|PA|NJ|Reading|OTHER|Portsmouth|West Chester'),
                  'location'] = 'Northeast United States'

dice.loc[dice['location'].str.contains('MI|OH|IN|IL|Il|WI|MN|IA|Ia|MO|Mo|KS|NE|ND|SD|Farmington|Lake County|Ashland'),
                  'location'] = 'Midwest United States'

dice.loc[dice['location'].str.contains('DE|MD|Md|DC|WV|VA|NC|SC|Anderson|KY|Ky|TN|GA|Ga|MS|AL|AR|OK|TX|Stafford|Austin|LA|La|FL|Boca Raton|South'),
                  'location'] = 'Southern United States'

dice.loc[dice['location'].str.contains('MT|WY|CO|Co|NM|AZ|UT|ID|Id|NV|CA|Ca|OR|WA|AK|HI|Hi|Pasadena|Redmond|Greenwood Village|San Francisco'),
                  'location'] = 'Western United States'

dice.loc[dice['location'].str.contains('Hyderabad|London|Bangalore|Dublin|Taguig City|ON|BC|Windsor'),
                  'location'] = 'International'

dice.loc[dice['location'].str.contains('STATE|City|Satellite Office; North America-us-il-chicago; Nor|United States Of America'),
                  'location'] = 'Unknown'
dice.head(3)

# Creating a list of top 4 locations
keep_locations = dice['location'].value_counts()[:4].index.tolist()

# Only keeping the top 4 locations from list keep_locations
dice = dice[dice['location'].isin(keep_locations)]

dice['location'].value_counts()

## Cleaning job_type

# Droping 238 rows of data that are null for job_type.
dice.dropna(subset = ['job_type'], inplace=True)
dice.reset_index(drop=True, inplace=True)

### Catagorizing job_type

dice.loc[dice['job_type'].str.contains('Full Time|Full-time|per year|per'),
                  'job_type'] = 'Full Time'

dice.loc[dice['job_type'].str.contains('Contract|contract|C2H'),
                  'job_type'] = 'Contract'

dice.loc[dice['job_type'].str.contains('Part Time'),
                  'job_type'] = 'Part Time'

dice.loc[dice['job_type'].str.contains('Market related|Market|Negotiable'),
                  'job_type'] = 'Market Dependent'

dice.loc[dice['job_type'].str.contains('-|EXPERIENCE|RELOCATION|define'),
                  'job_type'] = 'Unknown'

dice.head(3)

dice['job_type'].value_counts()

## Cleaning sector

# renaming sector to be more descriptive
dice=dice.rename(columns = {'sector':'skills'})

# lowercasing skills
dice['skills']= dice['skills'].str.lower()

### Removing commas, backslashes and extra spaces

# making sure no extra spaces. help from : https://stackoverflow.com/questions/43071415/remove-multiple-blanks-in-dataframe
dice['skills'] = dice['skills'].replace('\s+', ' ', regex=True)

### Looking at skills string length

# creating a new column for skills string length
dice['skills_len'] = dice['skills'].str.split().map(lambda x: len(x))

dice[dice['skills_len'] ==2].head(3)

dice[dice['skills_len'] ==3].head(3)

We can see that many of the skills are actually just the title of the job or job type. These should be changed to an empty string.

### Cleaning skills of already described features

# if skill is just the job title replace with a empty string
dice.loc[(dice['skills'] ==  dice['job_title'].str.lower()), 'skills'] = ""

dice['skills'].value_counts().head(10)

# changing additional non-skills to empty strings
dice.loc[(dice['skills'] == 'Telecommuting not available Travel not required'.lower()), 'skills'] = ''
dice.loc[(dice['skills'] == 'Full Time'.lower()), 'skills'] = ''
dice.loc[(dice['skills'] == 'Contract W2'.lower()), 'skills'] = ''
dice.loc[(dice['skills'] == 'tad pgs inc. specializes in delivering secure reliable and rapidly implemented workforce solutions to the u.s. federal marketplace including u.s. government agencies and their prime contractors. wi'), 'skills'] = ''
dice.loc[(dice['skills'] == 'see job description'), 'skills'] = ''
dice.loc[(dice['skills'] == 'please refer to job description'), 'skills'] = ''
dice.loc[(dice['skills'] == '(see job description)'), 'skills'] = ''
dice.loc[(dice['skills'] == 'refer to job description'), 'skills'] = ''

### Dealing with duplicate words

# removing duplicate words in a row
# help from:https://stackoverflow.com/questions/47316783/python-dataframe-remove-duplicate-words-in-the-same-cell-within-a-column-in-pyt
dice['skills'] = dice['skills'].str.replace(r'\b(\w+)(\s+\1)+\b', r'\1')

# some postings had repeat word after -
# mapping split_skills to whole skills row
dice['skills'] = dice['skills'].map(split_skills)

# Remove common phrases
dice['job_description'] = dice['job_description'].map(rem_fluff)
dice['skills'] = dice['skills'].map(rem_fluff)

# Correct multiple spaces (including \t and \n)
dice['job_description'] = dice['job_description'].map(lambda x: re.sub('\s+',' ', x))
dice['skills'] = dice['skills'].map(lambda x: re.sub('\s+',' ', x))

# Final DataFrame

dice.reset_index(drop=True, inplace=True)

dice

# exporting final dataframe to csv
dice.to_csv('../data/job_postings.csv', index=False)
