==> code/02-process-linkedin-data.ipynb <==
# Process LinkedIn Data

import pandas as pd
import re

import httpimport
url = 'https://raw.githubusercontent.com/zach-brown-18/class-toolkit/main/eda/'
with httpimport.remote_repo(['cleaning', 'columns'], url):
    import cleaning as c
    import columns as cols

def remove_unwanted_characters(s):
    s = re.sub('\*', ' ', s)
    s = re.sub('/', ' ', s)
    s = re.sub('_', ' ', s)
    s = re.sub('\?', ' ', s)
    s = re.sub('%', ' ', s)
    s = re.sub('@', ' ', s)
    s = re.sub('#', ' ', s)
    s = re.sub('!', ' ', s)
    s = re.sub(',', ' ', s)
    s = re.sub('\+', ' ', s)
    s = re.sub('&', ' ', s)
    s = re.sub('\d', ' ', s)
    return s

def clean_linkedin_data(df):
    for col in df:
        df[col] = df[col].map(remove_unwanted_characters)
        df[col] = df[col].map(c.remove_special_chars)
        df[col] = df[col].str.lower()
    return df

def import_file(file_name, url):
    try:
        df = pd.read_csv(url + file_name)
        return df
    except FileNotFoundError:
        print(f"Missing {file_name} file.")

def linkedin_data(folder_name):
    # Loading files
    url = f'../data/linkedin/{folder_name}/'
    
    skills_data = import_file('Skills.csv', url)
    positions_data = import_file('Positions.csv', url)
    profile_data = import_file('Profile.csv', url)
    education_data = import_file('Education.csv', url)
    project_data = import_file('Projects.csv', url)
    certification_data = import_file('Certifications.csv', url)
    
    # Extracting information
    try:
        full_name = profile_data['First Name'] + " " + profile_data['Last Name']
        summary = profile_data['Summary']
    except:
        print("Profile columns could not be created.")

    try:
        position_list = positions_data['Title'].values.tolist()
        position_string = ', '.join([str(position) for position in position_list])
    except:
        print("Positions columns could not be created.")
    
    try:
        skills_list = skills_data.values.tolist()
        skills_string = ', '.join([str(skill) for skill in skills_list])
        skills_string = skills_string.replace("['", "").replace("']", "")
    except:
        print("Skills columns could not be created.")

    try:
        education_list = education_data['Degree Name'].values.tolist()
        education_string = ', '.join([str(education) for education in education_list])
    except:
        print("Education columns could not be created.")

    try:
        jobtype_list = jobpref_data['Preferred Job Types'].values.tolist()
        jobtype_string = ', '.join([str(jtype) for jtype in jobtype_list])
        jobtype_string = jobtype_string.replace(" | ", ", ")
        
        prefjob_list = jobpref_data['Job Titles'].values.tolist()
        prefjob_string = ', '.join([str(pjob) for pjob in prefjob_list])
        prefjob_string = prefjob_string.replace(" | ", ", ")
    except:
        print("Job preference columns could not be created.")
    
    try:
        project_title_list = project_data['Title'].values.tolist()
        project_title_string = ', '.join([str(project).strip() for project in project_title_list])
        project_desc_list = project_data['Description'].values.tolist()
        project_desc_string = ', '.join([str(project).strip() for project in project_desc_list])
        project_desc_string = ', ' + project_desc_string
        project_list = project_title_string + project_desc_string
    except:
        print("Projects column could not be created.")
        
    try:
        certification_list = certification_data['Name'].values.tolist()
        certification_string = ', '.join([str(cert).strip() for cert in certification_list])
    except:
        print("Certification column could not be created.")
    
    # Building dataframe
    data_df = {}
    
    try: data_df['Name'] = full_name
    except: print('Missing name.')
    
    try: data_df['Titles'] = position_string
    except: print('Missing titles.')
    
    try: data_df['Skills'] = skills_string
    except: print('Missing skills.')
    
    try: data_df['Summary'] = summary
    except: print('Missing summary.')

    try: data_df['Education'] = education_string
    except: print('Missing education.')

    try: data_df['Job Type Preference'] = jobtype_string
    except: print('Missing job type preference.')

    try: data_df['Preferred Job Title'] = prefjob_string
    except: print('Missing preferred job title.')
        
    try: data_df['Projects'] = project_list
    except: print('Missing projects.')
        
    try: data_df['Certifications'] = certification_string
    except: print('Missing certifications.')

    # Saving dataframe
    try:
        df = pd.DataFrame(data=data_df)
        df = clean_linkedin_data(df)
        df.to_csv(f'../data/linkedin/test-output/{folder_name}.csv', index=False)
    except:
        print('Failed to save dataframe.')
    
    return df

# Test it out here

df_zach = linkedin_data('Zach_LinkedInData_12-16-2020')

df_zach

df_nolan = linkedin_data('Nolan_LinkedInData_12-16-2020')

df_nolan

df_albert = linkedin_data('Albert_LinkedInData')

df_albert

df_ye = linkedin_data('Ye_LinkedInData')

df_ye
