==> code/03-CVec.ipynb <==
# CountVectorizer Model

### Code Overview
For the CountVectorizer Model, we will want to fit transform our job posting column 'text', which is a collection of job description words and job skill words. We will then transform the 'text' column in our user data which is the users titles, skills, summary, education, certifications, and projects columns. We decided to create a function that would import necessary data files and create a dataframe of user information to run in our recommendation system.

We then calculate a cosine similiarty score from the user linkedin data and the job posting dataset. The cosine similarity score is a measure of the simlarity between each job title (the 13 we created), and the users linkedin information. The higher the cosine similarity score, the more likely a job title is to be recommended. The output would be a dataframe that would order the similarity scores from highest to lowest next to the job title that was most likely to be recommended.

import pandas as pd
import numpy as np
import re
from sklearn.feature_extraction.text import TfidfVectorizer, CountVectorizer
from sklearn.metrics.pairwise import linear_kernel, cosine_similarity
from IPython.display import display_html

## Jobs data

jobs = pd.read_csv('../data/job_postings.csv')
jobs = jobs.drop(columns=['date_added', 'organization', 'skills_len', 'job_type'])
jobs.fillna('', inplace=True)
jobs['text'] = jobs['job_description'] + ' ' + jobs['skills']

## User data

def gather_profile_data(file_path):
    profile_data = pd.read_csv(file_path)
    profile_data['text'] = profile_data['Titles'] + ' ' \
                            + profile_data['Skills'] + ' ' \
                            + profile_data['Summary'] + ' ' \
                            + profile_data['Education']
    try: profile_data['text'] += ' ' + profile_data['Certifications']
    except: pass
    
    try: profile_data['text'] += ' ' + profile_data['Projects']
    except: pass
    
    return profile_data

# Reading in linkedin profile data.
profile_data_zach = gather_profile_data('../data/linkedin/test-output/Zach_LinkedInData_12-16-2020.csv')
profile_data_nolan = gather_profile_data('../data/linkedin/test-output/Nolan_LinkedInData_12-16-2020.csv')
profile_data_albert = gather_profile_data('../data/linkedin/test-output/Albert_LinkedInData.csv')
profile_data_ye = gather_profile_data('../data/linkedin/test-output/Ye_LinkedInData.csv')

## Make recommendations

def get_recommendations(vectorizer, user_data):
    # Fit transform on text data
    cvec_jobtext = vectorizer.fit_transform(jobs['text'])

    # Transforming user profile text
    user_cvec = vectorizer.transform(user_data['text'])

    # Calculating cosine similarity between users profile and job text
    cos_similar_user = map(lambda x: cosine_similarity(user_cvec, x), cvec_jobtext)

    # Results are a list of our cosine similarity scores
    results = list(cos_similar_user)

    # Create a dataframe using our results and job_titles
    new_df = pd.DataFrame(data = [results, jobs['job_title']]).T
    new_df = new_df.rename(columns = {0: "sim_score", 1 : "job_title"})

    # Column sim_score as float so that we can use groupby
    new_df['sim_score'] = new_df['sim_score'].astype(float)

    # Group by job_title, using count and mean to sort values
    new_df = new_df.groupby('job_title').mean().sort_values('sim_score', ascending=False)

    return new_df

# Instantiating countvectorizer
count_vectorizer = CountVectorizer()

# Instantiating countvectorizer with stopwords and min_df
count_vectorizer_stopwords = CountVectorizer(stop_words = "english", min_df = 10)

# Keep all words
zachs_recommendations = get_recommendations(count_vectorizer, profile_data_zach)
nolans_recommendations = get_recommendations(count_vectorizer, profile_data_nolan)
alberts_recommendations = get_recommendations(count_vectorizer, profile_data_albert)
yes_recommendations = get_recommendations(count_vectorizer, profile_data_ye)

# Remove stop words and min_df = 10
zachs_recommendations_stopwords = get_recommendations(count_vectorizer_stopwords, profile_data_zach)
nolans_recommendations_stopwords = get_recommendations(count_vectorizer_stopwords, profile_data_nolan)
alberts_recommendations_stopwords = get_recommendations(count_vectorizer_stopwords, profile_data_albert)
yes_recommendations_stopwords = get_recommendations(count_vectorizer_stopwords, profile_data_ye)

### Zach

# Credit for notebook styling: https://blog.softhints.com/display-two-pandas-dataframes-side-by-side-jupyter-notebook/
df1_styler = zachs_recommendations.style.set_table_attributes("style='display:inline'").set_caption('All Words')
df2_styler = zachs_recommendations_stopwords.style.set_table_attributes("style='display:inline'").set_caption('Stopwords Removed')

space = "\xa0" * 10
display_html(df1_styler._repr_html_() + space + df2_styler._repr_html_(), raw=True)

### Nolan

df1_styler = nolans_recommendations.style.set_table_attributes("style='display:inline'").set_caption('All Words')
df2_styler = nolans_recommendations_stopwords.style.set_table_attributes("style='display:inline'").set_caption('Stopwords Removed')

space = "\xa0" * 10
display_html(df1_styler._repr_html_() + space + df2_styler._repr_html_(), raw=True)

### Albert

df1_styler = alberts_recommendations.style.set_table_attributes("style='display:inline'").set_caption('All Words')
df2_styler = alberts_recommendations_stopwords.style.set_table_attributes("style='display:inline'").set_caption('Stopwords Removed')

space = "\xa0" * 10
display_html(df1_styler._repr_html_() + space + df2_styler._repr_html_(), raw=True)

### Ye

df1_styler = yes_recommendations.style.set_table_attributes("style='display:inline'").set_caption('All Words')
df2_styler = yes_recommendations_stopwords.style.set_table_attributes("style='display:inline'").set_caption('Stopwords Removed')

space = "\xa0" * 10
display_html(df1_styler._repr_html_() + space + df2_styler._repr_html_(), raw=True)

### CountVectorizer Results
Our countvectorizer model recommended similar positions across all of the group members, with Data Position scoring the highest for each of our linkedin profile data. When running our model on a profile that was not in the tech industry, the countvectorizer model was able to analyze the data and give us results that show that their profile was not formatted for a technical position. This means that our model can accurately predict whether or not a users profile information is technical, and/or non-technical.

The countvectorizer model performed how our group predicted, and with consistent results across all profiles it was tested on. In conclusion, our countvectorizer model was the most accurate model and had scores that accurately reflected our predictions before modeling even began. With more time and modifications to our jobs dataset, we feel comfortable making the decision to choose the countvectorizer model as the best performing model.
