==> code/03-KNN.ipynb <==
# KNN Model

## KNN description:
Goes through the same steps as count vectorizer, but gathers the 75 most closely related job postings.
Calculates the frequency of each job title in those 75 most related jobs.
Job recommendation is then based on the most frequent job title of the 75 most closely related posts to the LinkedIn profile.

import pandas as pd
import numpy as np
import re
from sklearn.feature_extraction.text import TfidfVectorizer, CountVectorizer
from sklearn.metrics.pairwise import linear_kernel, cosine_similarity
from sklearn.neighbors import NearestNeighbors
from IPython.display import display_html

## Jobs data

jobs = pd.read_csv('../data/job_postings.csv')
jobs = jobs.drop(columns=['date_added', 'organization', 'skills_len', 'job_type'])
jobs.fillna('', inplace=True)
jobs['text'] = jobs['job_description'] + ' ' + jobs['skills']

jobs

## User data

def gather_profile_data(file_path):
    profile_data = pd.read_csv(file_path)
    profile_data['text'] = profile_data['Titles'] + ' ' \
                            + profile_data['Skills'] + ' ' \
                            + profile_data['Summary'] + ' ' \
                            + profile_data['Education']
    try: profile_data['text'] += ' ' + profile_data['Certifications']
    except: pass
    
    try: profile_data['text'] += ' ' + profile_data['Projects']
    except: pass
    
    return profile_data

# Reading in linkedin profile data.
profile_data_zach = gather_profile_data('../data/linkedin/test-output/Zach_LinkedInData_12-16-2020.csv')
profile_data_nolan = gather_profile_data('../data/linkedin/test-output/Nolan_LinkedInData_12-16-2020.csv')
profile_data_albert = gather_profile_data('../data/linkedin/test-output/Albert_LinkedInData.csv')
profile_data_ye = gather_profile_data('../data/linkedin/test-output/Ye_LinkedInData.csv')

## Make recommendations

def get_recommendations(vectorizer, tfidf_jobtext, user_data):
    # Transforming user profile text
    user_tfidf = vectorizer.transform(user_data['text'])

    # Calculating KNN similarity between users profile and job text (top 100 most similar jobs)
    n_neighbors=75
    KNN = NearestNeighbors(n_neighbors=n_neighbors, p=2, metric='cosine', algorithm = 'brute')
    KNN.fit(tfidf_jobtext)
    NNs = KNN.kneighbors(user_tfidf, return_distance=True)

    # Finding indexs for n_neighbors most similar jobs
    index = list(NNs[1][0][1:]) # indexs for top jobs
    final_jobs = jobs.loc[index] 
    
    # Create a dataframe using our job title counts(top 10)
    pos_df = pd.DataFrame(final_jobs['job_title'].value_counts()[:10])
    
    # Renaming job_title
    pos_df['Job Count'] = pos_df['job_title']
    pos_df.drop(columns='job_title', inplace= True)
    
    # Creating column for percent of jobs matched
    pos_df['Job Match %'] = pos_df['Job Count']/n_neighbors
    
    return pos_df

## Specific Recommendations

# Instantiating Tfidfvectorizer
tfidf_vectorizer = TfidfVectorizer(stop_words='english', ngram_range=(1,2))
tfidf_jobtext = tfidf_vectorizer.fit_transform(jobs['text'])

# Calculate recommendations
nolans_recommendations = get_recommendations(tfidf_vectorizer, tfidf_jobtext, profile_data_nolan)
zachs_recommendations = get_recommendations(tfidf_vectorizer, tfidf_jobtext, profile_data_zach)
alberts_recommendations = get_recommendations(tfidf_vectorizer, tfidf_jobtext, profile_data_albert)
yes_recommendations = get_recommendations(tfidf_vectorizer, tfidf_jobtext, profile_data_ye)

# Credit for notebook styling: https://blog.softhints.com/display-two-pandas-dataframes-side-by-side-jupyter-notebook/
df1_styler = zachs_recommendations.style.set_table_attributes("style='display:inline'").set_caption('Zach')
df2_styler = nolans_recommendations.style.set_table_attributes("style='display:inline'").set_caption('Nolan')
df3_styler = alberts_recommendations.style.set_table_attributes("style='display:inline'").set_caption('Albert')
df4_styler = yes_recommendations.style.set_table_attributes("style='display:inline'").set_caption('Ye')

space = "\xa0" * 50
display_html(df1_styler._repr_html_() + space + df2_styler._repr_html_() + space + df3_styler._repr_html_() + space + df4_styler._repr_html_(), raw=True)

## KNN Conclusion
Most common job matches appear to line up with what we would expect. Albert, Nolan and Zach were all recommended Data positions while Ye was recommended a Developer position.
Job match % may not be the most useful metric to compare other models to but it does give a sense of how well we can trust our given job output. 
KNN is also influenced by the balance of job titles and data positions was by far one of the least common. Yet, we still got recommended data positions which shows some strength to this model.
