==> code/03-Spacy.ipynb <==
# SpaCy

### Overview
Here we use a pre-trained model from the SpaCy package ("en_core_web_lg") to build a recommender system. The SpaCy model is able to process word context, meaning, and sentence structure when calculating similarity scores. SpaCy allows for a deeper analysis into each body of text as compared to the CountVectorizer, Tf-Idf Vectorizer and KNN models. We had high expectations for this model, however as you will see, the model fails to make accurate recommendations in this project.

import pandas as pd
import numpy as np
import re
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.metrics.pairwise import linear_kernel
import spacy
from IPython.display import display_html
from sklearn.feature_extraction import text

---

jobs = pd.read_csv('../data/job_postings.csv')
jobs = jobs.drop(columns=['date_added', 'organization', 'skills_len', 'job_type'])

jobs['job_title'].unique()

jobs.head(5)

jobs.isna().sum()

# Replace nan's with empty string
jobs.fillna('', inplace=True)

# Combine job description and text into a single column
jobs['text'] = jobs['job_description'] + ' ' + jobs['skills']

def remove_stopwords(lst):
    for word in lst:
        if word in text.ENGLISH_STOP_WORDS:
            lst.remove(word)
    return lst

text_split = jobs['text'].map(lambda x: x.split())
text_split_no_stopwords = text_split.map(remove_stopwords)
text_no_stopwords = text_split_no_stopwords.map(lambda x: ' '.join(x))

jobs['text'] = text_no_stopwords

# Spacy
https://spacy.io/usage/spacy-101

# Load one of the larger models for a better similarity score
nlp = spacy.load("en_core_web_lg")

# Create spacy documents for each job post
titles_and_docs = jobs[['job_title']].copy()
titles_and_docs['doc'] = jobs['text'].map(nlp)

titles_and_docs

def gather_profile_data(file_path):
    profile_data = pd.read_csv(file_path)
    profile_data['text'] = profile_data['Titles'] + ' ' \
                            + profile_data['Skills'] + ' ' \
                            + profile_data['Summary'] + ' ' \
                            + profile_data['Education']
    try: profile_data['text'] += ' ' + profile_data['Certifications']
    except: pass
    
    try: profile_data['text'] += ' ' + profile_data['Projects']
    except: pass
    
    return profile_data

def get_recommendations(profile_data):
    # Create nlp doc from profile
    profile_text = profile_data['text'][0]
    profile_doc = nlp(profile_text)
    
    # Calculate scores
    scores = jobs[['job_title']].copy()
    scores['sim_score'] = titles_and_docs['doc'].map(lambda x: x.similarity(profile_doc))
    
    return scores

# Reading in linkedin profile data.
profile_data_zach = gather_profile_data('../data/linkedin/test-output/Zach_LinkedInData_12-16-2020.csv')
profile_data_nolan = gather_profile_data('../data/linkedin/test-output/Nolan_LinkedInData_12-16-2020.csv')
profile_data_albert = gather_profile_data('../data/linkedin/test-output/Albert_LinkedInData.csv')
profile_data_ye = gather_profile_data('../data/linkedin/test-output/Ye_LinkedInData.csv')

# How much text is in each of our profiles?
print('Zach:', len(profile_data_zach['text'][0]))
print('Nolan:', len(profile_data_nolan['text'][0]))
print('Albert:', len(profile_data_albert['text'][0]))
print('Ye:', len(profile_data_ye['text'][0]))

# Calculate scores
zach_scores = get_recommendations(profile_data_zach)
nolan_scores = get_recommendations(profile_data_nolan)
albert_scores = get_recommendations(profile_data_albert)
ye_scores = get_recommendations(profile_data_ye)

# Group by job title
zachs_recommendations = zach_scores.groupby('job_title').mean().sort_values('sim_score', ascending=False)
nolans_recommendations = nolan_scores.groupby('job_title').mean().sort_values('sim_score', ascending=False)
alberts_recommendations = albert_scores.groupby('job_title').mean().sort_values('sim_score', ascending=False)
yes_recommendations = ye_scores.groupby('job_title').mean().sort_values('sim_score', ascending=False)

# Credit for notebook styling: https://blog.softhints.com/display-two-pandas-dataframes-side-by-side-jupyter-notebook/
df1_styler = zachs_recommendations.style.set_table_attributes("style='display:inline'").set_caption('Zach')
df2_styler = nolans_recommendations.style.set_table_attributes("style='display:inline'").set_caption('Nolan')
df3_styler = alberts_recommendations.style.set_table_attributes("style='display:inline'").set_caption('Albert')
df4_styler = yes_recommendations.style.set_table_attributes("style='display:inline'").set_caption('Ye')

space = "\xa0" * 5
display_html(df1_styler._repr_html_() + space + df2_styler._repr_html_() + space + df3_styler._repr_html_() + space + df4_styler._repr_html_(), raw=True)

### SpaCy Conclusion
As shown in the results above, the similarity scores for each LinkedIn profile hardly differ from the highest recommended title to the lowest recommended title. This is particularly apparent in Zach's scores, where the highest recommended position, Analyst, is only .0255 greater than the lowest recommended position of Consulting. We conclude from these results that this recommender is significantly effected by the noise, i.e. common words, in the dataset, and thus fails to provide accurate recommendations. This approach shows promise in future applications, as we fine tune the data cleaning process, however in the current state we decide not to use SpaCy.
